+++ src/haddock/core/supported_molecules.py
@@ -24,6 +24,8 @@ not done automatically on purpose.
 * ``solvent-allhdg5-4.top``
 * ``shape.top``
 * ``cofactors.top``
+* ``lipids.top``
+
 
 DNA-related files are all concatenated into a single DNA-supported
 residues datastructure. The same occurs for protein residues (natural or
@@ -44,6 +46,7 @@ supported residue names:
 * :py:data:`supported_solvents_resnames`
 * :py:data:`supported_shape`
 * :py:data:`supported_cofactors_resnames`
+* :py:data:`supported_lipids_resnames`
 
 * :py:data:`supported_ATOM` (contains residues and nucleic acids)
 * :py:data:`supported_HETATM` (everything not contained in ATOM)
@@ -205,7 +208,7 @@ def _read_residues_from_top_file(
 
     # all lines are striped before looping
     for line in map(str.strip, lines):  # it is important to strip spaces
-        if line.startswith(('RESI', 'residue')):
+        if line.startswith(('RESI', 'residue', 'Residue')):
             in_residue = True  # defines we start a new residue
 
             # extracts the resname from the residue line
@@ -288,6 +291,7 @@ def read_supported_residues(
     * ``solvent-allhdg5-4.top``
     * ``shape.top``
     * ``cofactors.top``
+    * ``lipids.top``
 
     You need to edit this function to account for any additional
     ``.top`` file that is added to the ``source_path`` on top of the
@@ -313,6 +317,7 @@ def read_supported_residues(
     solvent_top = Path(source_path, "solvent-allhdg5-4.top")
     shape_top = Path(source_path, "shape.top")
     cofactors_top = Path(source_path, "cofactors.top")
+    lipids_top = Path(source_path, "lipids.top")
 
     # supported Residues (tuple of namedtuples)
     supported_carbohydrates = read_residues_from_top_file(carbo_top)
@@ -352,6 +357,7 @@ def read_supported_residues(
     supported_solvents = read_residues_from_top_file(solvent_top)
     supported_shape = read_residues_from_top_file(shape_top)
     supported_cofactors = read_residues_from_top_file(cofactors_top)
+    supported_lipids = read_residues_from_top_file(lipids_top)
 
     # supported resnames
     supported_carbo_resnames = get_resnames(supported_carbohydrates)
@@ -364,6 +370,7 @@ def read_supported_residues(
     supported_solvents_resnames = get_resnames(supported_solvents)
     supported_shape_resnames = get_resnames(supported_shape)
     supported_cofactors_resnames = get_resnames(supported_cofactors)
+    supported_lipids_resnames = get_resnames(supported_lipids)
 
     # other attributes
     for ion in supported_single_ions:
@@ -387,6 +394,7 @@ def read_supported_residues(
         supported_solvents_resnames,
         supported_shape_resnames,
         supported_cofactors_resnames,
+        supported_lipids_resnames,
         )
 
 
@@ -402,7 +410,8 @@ _supported_carbo_resnames, \
     _supported_aminoacids_resnames, \
     _supported_solvents_resnames, \
     _supported_shape_resnames, \
-    _supported_cofactors_resnames = read_supported_residues(toppar_path)
+    _supported_cofactors_resnames, \
+    _supported_lipids_resnames = read_supported_residues(toppar_path)
 
 # render docstrings
 supported_carbo_resnames = set(_supported_carbo_resnames)
@@ -450,6 +459,9 @@ supported_shape_resnames = set(_supported_shape_resnames)
 supported_cofactors_resnames = set(_supported_cofactors_resnames)
 """Supported cofactors."""
 
+supported_lipids_resnames = set(_supported_lipids_resnames)
+"""Supported lipids residues names."""
+
 #
 # Residues that must be set as ATOM
 supported_ATOM = set(it.chain(
@@ -473,6 +485,7 @@ supported_non_ions = set(it.chain(
     supported_solvents_resnames,
     supported_shape_resnames,
     supported_cofactors_resnames,
+    supported_lipids_resnames,
     ))
 
 # Residues that must be set as HETATM
@@ -484,6 +497,7 @@ supported_HETATM = set(it.chain(
     supported_multiatom_ions_resnames,
     supported_solvents_resnames,
     supported_cofactors_resnames,
+    supported_lipids_resnames,
     ))
 """
 Supported ``HETATM`` residues.